+++ client/waatea2_client/lib/models/setavailability_model.dart
@@ -2,7 +2,7 @@ class SetAvailabilityModel {
   String games;
   final String date;
   int state;
-  String avail_id;
+  String availId;
   final int dayofyear;
   final String season;
 
@@ -10,7 +10,7 @@ class SetAvailabilityModel {
       {required this.games,
       required this.date,
       required this.state,
-      required this.avail_id,
+      required this.availId,
       required this.dayofyear,
       required this.season});
 }