+++ client/waatea2_client/lib/models/user_model.dart
@@ -5,14 +5,14 @@ class UserModel {
   final String name;
   final String email;
   final UserProfileModel profile;
-  final int attendance_percentage;
+  final int attendancePercentage;
 
   UserModel(
       {required this.pk,
       required this.name,
       required this.email,
       required this.profile,
-      required this.attendance_percentage});
+      required this.attendancePercentage});
 
   factory UserModel.fromJson(Map<String, dynamic> json) {
     return UserModel(
@@ -20,7 +20,7 @@ class UserModel {
         name: json['name'],
         email: json['email'],
         profile: UserProfileModel.fromJson(json['profile']),
-        attendance_percentage: json['attendance_percentage']);
+        attendancePercentage: json['attendance_percentage']);
   }
 
   Map<String, dynamic> toJson() => {