+++ client/waatea2_client/lib/widgets/showplayerattendance.dart
@@ -22,12 +22,12 @@ class AttendedViewModel {
 }
 
 class _ShowPlayerAttendanceState extends State<ShowPlayerAttendance> {
-  late Future<List<AttendedViewModel>> train_attend_list;
+  late Future<List<AttendedViewModel>> trainAttendList;
 
   @override
   void initState() {
     super.initState();
-    train_attend_list = fetchTrainings();
+    trainAttendList = fetchTrainings();
   }
 
   Future<List<AttendedViewModel>> fetchTrainings() async {
@@ -55,7 +55,7 @@ class _ShowPlayerAttendanceState extends State<ShowPlayerAttendance> {
       backgroundColor: Colors.transparent,
       body: Center(
         child: FutureBuilder<List<AttendedViewModel>>(
-          future: train_attend_list,
+          future: trainAttendList,
           builder: (BuildContext context, AsyncSnapshot snapshot) {
             // By default, show a loading spinner.
             if (!snapshot.hasData)